backend/requirements.txt
# Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Embeddings / HTTP
requests>=2.31.0
sentence-transformers>=2.2.0

# Vector Stores
faiss-cpu>=1.7.4           # GPU version install manually
chromadb>=0.4.24
numpy>=1.24.0

# Document Handling
PyPDF2>=3.0.0
tiktoken>=0.5.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
pandas>=2.0.0

# Optional nice-to-haves
jinja2>=3.1.2
python-multipart>=0.0.7